Completely remove non cookie-based authentication mechanisms
It seems these weren't working anyway, due to google captcha, and the
cookie-based method is by far the most robust one, plus it adds some
additional security, not having to deal with passwords at all, and
simplifies the code.
Additionally, this makes the initial setup instructions clearer by
having only one method.

Overall, I would argue that logging in is not this script's role,
especially if it starts involving captchas. Feel free to create a
wrapper, or reuse the scrapped code to obtain the _simple_auth cookie
if you feel so inclined :)

Please open an issue if you were still using one of these methods, or
disagree with the reasoning.

This was shortly discussed in #5.