Actually remove MD5 file
Previously os.remove was called before HumbleHash.remove_md5file. However,
remove_md5file does nothing if the file (full_filename) does not exist.
(See hb-downloader/humble_api/humble_hash.py ).

So the md5 file was never actually removed and the call to remove_md5file did
nothing.

This lead to the problem that corrupt files were always re-downloaded again
and again without updating the corresponding md5 file. (The md5 file was
never deleted and it is assumed that existing md5 files are correct and
up-to-date.)

With this commit, the md5 file is actually deleted (and recreated when
necessary). This way there are no more out-of-date md5 files.